add detail/serializer.hpp
